Definitivo v1 API
